If execute times out, handle possible cancel timeout.
execute was fixed to close the cursor if it timed out, to prevent a
subsequent 'query in flight' error.  However, the test assumes that the
cancel will not time out.  This patch introduces a test for the case
where the cancel *also* times out, and fixes it by ignoring the cancel
timeout in the execute timeout error recovery.