Remove callbacks from a bad file descriptor immediately
Previously, the callback removal was postponed until .close() call, and
sometimes this .close() call cleared a file descriptor from a callback set
by another connection object (if the file desriptor was closed by psycopg2 when
an error occurred, and then opened again for a new connection).

The effect of this was, when making tens of parallel connections, an error in
one of them made another time out.

Now, if a psycopg2 error occurs, and the file descriptor has gone bad, the
callback is removed immediately, and the aiopg connection object 'forgets'
about this file descriptor (self._fileno = None), and doesn't try to remove
callbacks in .close().

If there were no errors, callbacks are still removed during .close() call.